add tls data and rpmalloc

+++ sources/sysapi/src/bezos/syscall_vmem.c
@@ -3,17 +3,17 @@
 #include <bezos/private.h>
 
 OsStatus OsAddressSpaceCreate(struct OsAddressSpaceCreateInfo CreateInfo, OsAddressSpaceHandle *OutHandle) {
-    struct OsCallResult result = OsSystemCall(eOsCallAddressSpaceCreate, (uintptr_t)&CreateInfo, 0, 0, 0);
+    struct OsCallResult result = OsSystemCall(eOsCallAddressSpaceCreate, (uint64_t)&CreateInfo, 0, 0, 0);
     *OutHandle = (OsAddressSpaceHandle)result.Value;
     return result.Status;
 }
 
 OsStatus OsAddressSpaceDestroy(OsAddressSpaceHandle Handle) {
-    struct OsCallResult result = OsSystemCall(eOsCallAddressSpaceDestroy, (uintptr_t)Handle, 0, 0, 0);
+    struct OsCallResult result = OsSystemCall(eOsCallAddressSpaceDestroy, (uint64_t)Handle, 0, 0, 0);
     return result.Status;
 }
 
 OsStatus OsAddressSpaceStat(OsAddressSpaceHandle Handle, struct OsAddressSpaceInfo *OutStat) {
-    struct OsCallResult result = OsSystemCall(eOsCallAddressSpaceStat, (uintptr_t)Handle, (uintptr_t)OutStat, 0, 0);
+    struct OsCallResult result = OsSystemCall(eOsCallAddressSpaceStat, (uint64_t)Handle, (uint64_t)OutStat, 0, 0);
     return result.Status;
 }
